feat: data_files expected location changed to name of the package
The location of the constraints file upon installation can be manually
configured and it was assumed that the folder name was the same as the
name of the top level package. It is more likely that the name will correspond
to the name of the python package.
Removed global constraints variable since we the script should be more
generic and not assume that all wheels have the same constraints