feat: determine whether an upgrade is necessary
There are two possible ways of determining whether an upgrade is
available. The first is to manually create a backup venv and
install a dependency in venv with `--upgrade` and then `pip check`.
However, running pip install to determine whether an upgrade is
available is expensive in backup venv is expensive.

The second approach is to rely on pip tooling that determines
whether an upgrade is available. However, to get the list of available
upgrades, we query the index.html page to get dependency candidates.
We can parse the html page to get dependency candidates,
as the PEP that proposes HTML standardization for packages
is finished and implemented in PEP 503 - https://peps.python.org/pep-0503/
Rely on lxml to parse the HTML contents into a tree.

After dependency candidates are parsed from anchor tags, we use
`packaging` `SpecifierSet` to filter out only the versions that are
compatible to the specifier. For examples, please see the docstring in
`_filter_versions`.