refact: rely on vendored pip packages
Instead of adding dependencies that are necessary/required to parse
wheels, parse requirements or make requests, pip vendors the
dependencies that are part of the pip package. Vendored packages MUST
NOT be modified except as required to successfully vendor them - [1].
This means that we delegate all package-handling logic to packages that
are bundled together with pip. The added benefit of using vendored
packages is that we don't have to manage them ourselves, depending on
which pip version gets installed.
Refactor upgrade script to use vendored `requests` and `packaging`.
This should hopefully dramatically simplify the upgrade-python-package
issue with adding more dependencies to the script.

[1] - https://pip.pypa.io/en/stable/development/vendoring-policy/