refact(upgrade): determine compatible local wheel version
Move `_filter_versions` function from venv manager to the upgrade
-python-package script. We should now be able to find the most
compatible local wheel given a set of version requirements. An example
would be if the version specifier set is ~=2.0.0, then the most
compatible version of a package should be ==2.0.2 for local wheels. This
was previously not the case, and resolves issues with upgrading local
dependencies.