feat(managevenv): add blue/green deployment flag
We do blue/green deployment of virtual environments. If blue/green
deployment is enabled, we expect the following venv directory structure:
- <ENVS_HOME_DIR>
   - dependency~=2.0.0_blue
   - dependency~=2.0.0_green

Management of switching the blue and green venvs is managed outside of
the script scope. If blue/green flag is disabled, we switch the upgraded_green
to venv _blue automatically.