fix(test): move `sys.executable` default into function
Testing framework patches the executable path to a test venv. Since
defaulting to sys.executable in function signature uses the default
global Python, rely on specifying a custom executable in function body
instead. Fix tests to not rely on success, since success is determined
by something being installed. Should not matter for test cases since we
are asserting the state of the venv.