test(managevenv): specify upgrade-python script version externally
We should now be able to specify an exact version of
upgrade-python-package script to be used in the new venvs. This allows
us to pin venvs to use a specific version of upgrade script, instead of
always using the latest one. Reason for pinning is that we don't want to
immediately transition to new upgrade-python-package script in venvs,
since a new release could potentially break the venvs.

This also enables us to patch the testing framework so venv management
tests should now work.